--- eventkit_cloud/ui/static/ui/app/actions/mapToolActionTypes.js
@@ -1,8 +1,0 @@
-const types = {
-    FILE_PROCESSING: 'FILE_PROCESSING',
-    FILE_PROCESSED: 'FILE_PROCESSED',
-    FILE_ERROR: 'FILE_ERROR',
-    FILE_RESET: 'FILE_RESET',
-};
-
-export default types;
