+++ eventkit_cloud/ui/static/ui/app/components/CreateDataPack/BreadcrumbStepper.js
@@ -15,8 +15,10 @@ import ExportAOI from './ExportAOI';
 import ExportInfo from './ExportInfo';
 import ExportSummary from './ExportSummary';
 import { flattenFeatureCollection } from '../../utils/mapUtils';
-import { getProviders, stepperNextDisabled,
-    submitJob, clearAoiInfo, clearExportInfo, clearJobInfo, getFormats } from '../../actions/exportsActions';
+import { stepperNextDisabled,
+    submitJob, clearAoiInfo, clearExportInfo, clearJobInfo } from '../../actions/exportsActions';
+import { getFormats } from '../../actions/formatActions';
+import { getProviders } from '../../actions/providerActions';
 import { getNotifications, getNotificationsUnreadCount } from '../../actions/notificationsActions';
 import BaseDialog from '../Dialog/BaseDialog';
 import ConfirmDialog from '../Dialog/ConfirmDialog';
