+++ eventkit_cloud/ui/static/ui/app/reducers/userActivityReducer.js
@@ -1,7 +1,15 @@
-import initialState from './initialState';
-import types from '../actions/actionTypes';
+import { types } from '../actions/userActivityActions';
 
-export function userActivityReducer(state = initialState.userActivity, action) {
+export const initialState = {
+    viewedJobs: {
+        fetching: false,
+        fetched: false,
+        viewedJobs: [],
+        error: null,
+    },
+};
+
+export function userActivityReducer(state = initialState, action) {
     switch (action.type) {
         case types.FETCHING_VIEWED_JOBS:
             return {
@@ -40,7 +48,7 @@ export function userActivityReducer(state = initialState.userActivity, action) {
                 },
             };
         case types.USER_LOGGED_OUT:
-            return initialState.userActivity;
+            return initialState;
         default:
             return state;
     }
